Move changing of map name to separate example
