+++ examples/basic/index.html
@@ -8,7 +8,6 @@ docs: >
   <li>Adding dynamic points to the GeoJSON layer</li>
   <li>Dynamically removing points from the GeoJSON layer</li>
   <li>Using the zoomSlider prop of the map component</li>
-  <li>Changing the map name alters the state.map.name value in the Redux store</li>
   <li>Using a mouse position control to show the latitude and longitude</li>
 
 ---
