+++ examples/mapname/index.html
@@ -1,0 +1,10 @@
+---
+layout: example.html
+title: Map Name Example
+shortdesc: Demonstrates how to change the name of a map.
+docs: >
+  <li>Changing the map name alters the state.map.name value in the Redux store</li>
+
+---
+<div id="features"></div>
+<div id="controls"></div>
